feat: error handling on controller added

+++ backend/controllers/goalsController.js
@@ -6,35 +6,39 @@ async function getGoals(req, res) {
   res.status(200).json(goals);
 }
 
-async function getGoal(req, res) {
+async function getGoal(req, res, next) {
   const { id } = req.params;
 
-  if(id.length !== 24) {
+  if (id.length !== 24) {
     res.status(400);
-    throw new Error('Invalid id provided');
+    const error = new Error('Invalid id provided');
+    return next(error);
   }
 
   const goal = await goalModel.findById(id);
 
-  if(!goal) {
+  if (!goal) {
     res.status(404);
-    throw new Error('Goal not found');
+    const error = new Error('Goal not found');
+    return next(error);
   }
 
   res.status(200).json(goal);
 }
 
-async function createGoal(req, res) {
+async function createGoal(req, res, next) {
   const body = req.body;
 
-  if(!body.text) {
+  if (!body.text) {
     res.status(400);
-    throw new Error('Text field is required');
+    const error = new Error('Text field is required');
+    return next(error);
   }
 
-  if(!body.completed) {
+  if (!body.completed) {
     res.status(400);
-    throw new Error('Completed field is required');
+    const error = new Error('Completed field is required');
+    return next(error);
   }
 
   const newGoal = await goalModel.create(body);
@@ -42,20 +46,22 @@ async function createGoal(req, res) {
   res.status(201).json(newGoal);
 }
 
-async function updateGoal(req, res) {
+async function updateGoal(req, res, next) {
   const { id } = req.params;
   const body = req.body;
 
-  if(id.length !== 24) {
+  if (id.length !== 24) {
     res.status(400);
-    throw new Error('Invalid id provided');
+    const error = new Error('Invalid id provided');
+    return next(error);
   }
 
   const goal = await goalModel.findById(id);
 
-  if(!goal) {
+  if (!goal) {
     res.status(404);
-    throw new Error('Goal not found');
+    const error = new Error('Goal not found');
+    return next(error);
   }
 
   const updatedGoal = await goalModel.findByIdAndUpdate(id, body, { new: true });
@@ -63,8 +69,13 @@ async function updateGoal(req, res) {
   res.status(200).json(updatedGoal);
 }
 
-async function deleteGoal(req, res) {
+async function deleteGoal(req, res, next) {
   const { id } = req.params;
+
+  if (id.length !== 24) {
+    const error = new Error('Invalid id provided');
+    return next(error);
+  }
 
   await goalModel.findByIdAndDelete(id);
 
@@ -75,6 +86,6 @@ module.exports = {
   getGoals,
   getGoal,
   createGoal,
-  updateGoal, 
+  updateGoal,
   deleteGoal,
 }
